+++ example/b.ts
@@ -1,28 +1,13 @@
 import { Elysia, t } from '../src'
 
-const subPlugin = new Elysia()
+new Elysia()
 	.guard({
-		as: 'global',
-		response: {
-			401: t.Literal('uma'),
-			403: t.Literal(':('),
+		query: t.Object({
+			id: t.Number()
+		})
+	})
+	.derive(({ query }) => {
+		return {
+			query
 		}
 	})
-
-const plugin = new Elysia()
-	.use(subPlugin)
-	.guard({
-		response: { 401: t.Number() }
-	})
-	.guard({
-		as: 'global',
-		response: { 418: t.Literal("I'm a teapot") }
-	})
-	.get('/', ({ error }) => error(401, 'uma'))
-	.get('/', ({ error }) => error(401, 1))
-	.get('/', ({ error }) => error(418, 'I\'m a teapot'))
-
-const app = new Elysia()
-	.use(plugin)
-	.get('/', ({ error }) => error(401, 'uma'))
-	.get('/', ({ error }) => error(418, 'I\'m a teapot'))